Filter events the env exposes in test snapshots to contracts to contract events (#1361)
### What
Filter events the env exposes in test snapshots to contracts to contract
events.

### Why
The main thing we need to test in test snapshots is that non-diagnostic
events are consistent and captured.

Diagnostic events are noise here.

### Notes
There's no good time to make this change.

Ideally the change is made at a major version because it's breaking.

But at a major version is the ideal time to avoid the change so that
people see only meaningful diffs on their test snapshots.