Fix breaking changes in latest stable and nightly rust (#1375)
### What

Pin the version of nightly in builds and CI, and make a few minor
changes.

### Why

The version is pinned because there's a bug in the next version of
nightly that we can't work around. The bug is:
- https://github.com/rust-lang/rust/issues/131643

The other minor changes that are to prepare for updating to the next
nightly and were required by it when using it.

### Known limitations

N/A