Add no-change check to build-fuzz and docs builds (#1383)
### What
Add no-change checks to the build-fuzz and docs builds, that errors if
there are any changes to the working directory after the builds run.

### Why
To detect files that weren't generated or updated relating to builds.
Also checks the lock file is up to date.

We could alternative use options on cargo to freeze the lock file, which
is another way to achieve same for that one file. However, this more
general check will check any files that changes as part of build, so
it's broader.