Add a link to migrating docs to home (#1389)
### What
Add a link to migrating docs to the home page in the docs, and update
other parts of the home page docs.

<img width="738" alt="Screenshot 2024-11-06 at 10 47 05 PM"
src="https://github.com/user-attachments/assets/77176926-385d-419a-ae24-c229a6ca35e9">


### Why
So that the migrating docs are easily discovered by anyone browsing the
Rust docs. To make it current.