Improve deploy docs (#1416)
### What
Remove the `let _ = deployer.deployed_address();` line from the two
existing deploy examples that are in the rust docs, and add an example
that discusses the `deployer.deployed_address()` in more detail.

<img width="948" alt="Screenshot 2024-12-13 at 11 07 22 pm"
src="https://github.com/user-attachments/assets/14735e0a-a32e-46ff-9e66-ff7ee21bdc6f"
/>

### Why
Plopping the `let _ = deployer.deployed_address();` into the middle of
an example that's about a specific scenario is distracting as to the
main purpose of the example. The docs can show how to use the function
in an example focused on it specifically.