Allow comparison of SDK types across Envs (#1436)
### What
Updates every type that is a handle to a type stored in the host so that
it is comparable with itself via a conversion to an `ScVal` even when
the values are across different `Env`s.

### Why
To make it much easier to perform differential testing.

Today it is not possible to compare values of an SDK type, when the
values are a handles to values stored on different hosts. This is
because the comparison of SDK types is supported by calling through to
the host. It makes sense for comparison to work this way because the
value is on the host side.

However when writing tests it is helpful to be able to compare SDK types
even if they belong to different Envs. The main use case is differential
tests, where two sequences of invokes should be comparable.

In tests the efficiency of the comparison is not important, so
converting to ScVal is an okay way to fall back to do the comparison. In
some ways this is a bit of a hack, but it's the least disruptive way of
supporting the comparison, and avoids introducing a ton of code to the
Env that would otherwise be out of place.

Talking to @graydon we have quite solid testing around the consistency
of ScVal and Val values, so converting the Vals to ScVal to do the
comparison is safe.

Close #1360