Switch to post-rust-1.84 wasm target wasm32v1-none (#1453)
Part of wrapping up bug
https://github.com/stellar/rs-soroban-sdk/issues/1428. Mostly this is
just:

- Switching rust version to 1.84 which supports `wasm32v1-none`
- Switching the SDK to build with that, not `wasm32-unknown-unknown`

But there are a couple minor subtle bits:

- Some of the crates only build native, not wasm, and we weren't
differentiating that in the Makefile right
- There _used_ to be a workaround for a mis-feature of `alloc` that
required `std` in order to register an OOM handler; this was actually
fixed in March 2023 (https://releases.rs/docs/1.68.0/ which contained
https://github.com/rust-lang/rust/pull/102318/) but the workaround was
already committed to the SDK by January 2023. Anyway, it's long since
fixed and we can just delete the workaround, which we _have_ to in this
PR, because `wasm32v1-none` doesn't ship with `std` at all.

Probably don't merge this until I have the other CI bits all switched
over to `wasm32v1-none` as described in #1428

---------

Co-authored-by: Leigh McCulloch <351529+leighmcculloch@users.noreply.github.com>