Add contractevent types for tokens and stellar asset contract (#1489)
### What

Add `contractevent` types for tokens and stellar asset contract.

Remove event publishing helpers.

### Why

With the addition of the `contractevent` macro in #1473 it makes sense
to use the macro to update the specs generated for the token and the
stellar asset contract such that their generated specs include the
events.

The `transfer` event added is represented with two event types:
- `Transfer`, which is the `single-value` data with an `amount`.
- `TransferMuxed`, which is the `map` data with a non-optional
`to_muxed_id`.

Two types are used because of the differing data format type, and the
contract event schema does not support multiple data format types to be
represented in a single type, to avoid complexity.

The reason that a non-optional `to_muxed_id` is used is because without
https://github.com/orgs/stellar/discussions/1750 the option none would
be stored as a void. In the future if that change occurs it may make
sense to update `TransferMuxed` to `Transfer2`, or similar, where-by it
represents some larger set of parameters that may grow over time.

The existing event publishing helpers are removed, rather than
deprecated, because the helpers build versions of the events that will
no longer be used by the stellar asset contract after CAP-67, versions
that include an admin topic. If the helpers remain, folks may continue
to call them and emit the topic. If the functions were updated, it would
be a breaking change anyway. It is more obvious to make the break by
removing them and encouraging use of the new events. Because the new
event types carry parameters as fields, it is very explicit that the
admin topic is gone.

Close https://github.com/stellar/rs-soroban-sdk/issues/1097
Close https://github.com/stellar/rs-soroban-sdk/issues/1488
 
### Merging

This change is intended to be merged to `main` after #1473.