Support ref fields on contractevent structs (#1499)
### What
Add support for reference fields in `contractevent` structs, allowing
borrowed values to be used in event fields.

  ### Why
When publishing events that are made up of user-defined-types, this
gives the developer the option of not copying the value into the event,
instead referencing it wherever it is. This may present some
optimisation opportunities over being forced to copy the value into the
struct. Nothing about this change requires the use of references, it
merely makes it allowed rather than disallowed.

Close #1491