Fix contractimpl macro composability with third party macros (#1510)
### What
Add a test vector that demonstrates a composability issue with third
party function macros. Apple the existing filtering of which attributes
get applied to the generated exported functions also to the generated
spec constant.

  ### Why
In #1427 I moved where the attributes were used as part of fixing
another bug, and inadvertently moved the expansion of the attributes to
the line before where the attributes are filtered. The filter ensures
that only a specific set of attributes get applied to the generated
code. The attributes still need to go through the filter for their use
on the spec constant and it was an error on my part not to also move the
filter.

Close #1509