Generate a seamless conversion from Address to MuxedAddress arguments in clients.
Also switch the token interface to use MuxedAddress, as this shouldn't be a breaking change anymore.

Besides removing the friction from the token interface users, this change corresponds to the protocol semantics of MuxedAddress - any contract that accepts MuxedAddress is going to be interoperable with the contracts that expect that it accepts just Address.