Fix reference type restrictions in map_type (#1518)
### What
Disallow reference types in map_type, offering up an error message
indicating that they are not supported, when map_type is instructed that
it is being used in a situation where reference types are not supported.

  ### Why
In #1499 I added support for ref types to map_type, to support ref types
on types that use the `contractevent` macro. When implementing I added
an `allow_ref` parameter that call sites used to indicate to the
map_type function if refs are supported. This is because while
contractevent types were getting ref support, all other use cases like
function args and contracttype types would not be. The reason that the
map_type function cares if they are supported or not is because it is
the central location where meaningful error messages about the type in
use are generated. If map_type doesn't error and provide a meaningful
error message, there will be other side-effect failures during
compilation that are just harder for the developer to reason about and
understand why they are occurring. Unfortunately when implementing ref
types into map_type I failed to actually add a guard to the function to
error if a ref type is detected in a situation that was not allowed.

This really begs the question why didn't tests catch this. The answer
being that unfortunately we don't have tests for compile failures
anymore. In the early days we did have a small suite of compile failure
tests that were effective at catching bugs like this, but because each
test triggered its own compiler it was extremely slow. I removed the
test suite in those early days because even with a small test suite the
cost to run it seemed out sized with the impact of bugs in this specific
area. We might want to consider bringing them back and running them on a
schedule or non-pr-blocking case. Separate to that this function should
have had some unit tests, it's perfectly suited for that, so I've
included a couple to start with and to cover the case being fixed.