Remove leading underscores from fn args in specs (#1513)
### What
Strip leading underscores from function argument names in the generated
contract spec.

  ### Why
Prevents implementation details from leaking into the spec and ensures
consistency across different implementations of the same trait.
Underscores are used as prefixes to argument names when a function
implementation doesn't make use of the argument. That's an
implementation detail, and surfacing that implementation detail into
specs causes different implementations of the same trait to be different
even though the interface is the same.

I debated about whether it should only strip a single underscore, or any
number of underscores. I went with it stripping all underscores because
I think it would be more surprising if `__arg` became `_arg` or stayed
as `__arg` when `_arg` becomes `arg`.

Close #1506