Deprecate old events publish method (#1524)
### What
Marked the publish method in the Events class as deprecated with a note
to use the #[contractevent] macro instead to define the event and
publish that value.

  ### Why
The #[contractevent] macro is the preferred way to define and publish
contract events as it will include the event into the contract spec.
Moving forward it is very much preferred way to define the event so that
downstream systems can generate bindings for contract events.