Add tests comparing token events to SAC events (#1530)
### What
Add tests comparing token events to SAC events. Fixes two bugs in how
the events were inconsistently defined.

  ### Why
We've noticed a couple bugs both now and in the long past of the token
interface in the SDK. There is currently nothing that validates ti is
correct and aligns with the Stellar Asset Contract's or SEP-41. This
change checks that the new events are consistent with the SAC's events,
minus the fact that the SAC includes an additional topic.