Lean into Transfer events with data as a map (#1534)
### What

Remove the `token::publish_transfer_event` function.

Rename Transfer to TransferLegacy.

Rename TransferMuxed to Transfer.

### Why

To encourage developers publishing the newer and easier to extend
Transfer event.

While it's true that the Stellar Asset Contract, for largely legacy
reasons, continues to publish a `transfer` event with the amount as the
data, there's no reason for the SDK to encourage other token
implementations to do the same.

Rather than giving developers a function to call to publish the transfer
event across these two different data forms, we should just point them
to publishing the newer form that is extendable.

### Merging

This PR is intended to merge to `main` after #1533. It is marked as a
draft PR until #1533 is merged.