Change how link sections are guaranteed to be included (#1543)
### What
Change how link sections are guaranteed to be included in contract
builds using a minimal length static export instead of a function
export.

  ### Why
Continues to ensure env-meta and SDK meta are preserved in the final
WASM even when a contract makes minimal use of the SDK. However, does so
without causing a function to be visible in the final binary.

Today a `_` function is visible in the final binary. From time to time
it is confusing to tool developers, who see the function and wonder what
it is. In general it is surprising and have always made me uncomfortable
as it looks like a hidden operation, even though it does nothing.

As @brson pointed out in
https://github.com/stellar/rs-soroban-sdk/issues/843#issuecomment-1407466584
a static global can be used instead. That static global doesn't actually
have to be of the data to embed, but can be a unit type, so minimally
sized, and its presence will cause the object file for the soroban-sdk
to be included in the link step and the impact to size is that file will
be smaller than it is today.

Close #843