Update copilot-instructions.md (#1550)
### What

Remove unused deps from Copilot instructions: rust-src, cargo-watch,
cargo-nextest.

### Why

They aren't required tools.

Telling Copilot to use rust-src is likely to send it on tangents we
don't want it to take. Recompiling rust-src and relying on features that
Rust provides only through rust-src would harm the developer experience
of folks using the soroban-sdk, requiring them to also use rust-src
which is a not a common Rust developer experience.

cargo-watch is a deprecated tool and not recommended to use. Unclear to
me why Copilot would benefit from it since it is more likely to benefit
from choosing to run appropriate test commands after each change, which
the instructions already detail.

cargo-nextest is not needed. It may speed up test runs, but we don't use
it in the Makefile and Copilot will benefit from using the test targets
in the Makefile to ensure that the test vectors build before tests.