Add mint event with mux support and other memo types to specs (#1555)
### What
Add a new Mint event with muxed address support. Add events to specs for
other memo types that can occur with payments from classic.

  ### Why
The mint event should be defined so that contract developers can choose
to emit a mint event to a destination with a muxed id.

The specs should contain all of the event structures that an ingester
could potentially see, including when other memo types are in use.

### Tests

It wasn't possible to add comparison tests for the mint muxed event, or
the mint memo string and bytes to_muxed_id types, to confirm the events
are identical. This is because the soroban-env doesn't allow us to
trigger those events because they are only generated by the classic
transaction processing.

For this reason @dmkozh @sisuresh please take great care to review the
events being added. I've written them based on CAP-67, but we've had
bugs in the past arise out of out-dated CAPs, so your help in validating
they match core's implementation is much appreciated.

Close https://github.com/stellar/rs-soroban-sdk/issues/1554