Break circular dependency between soroban-sdk and soroban-token-sdk. (#1558)
### What

Break circular dependency between soroban-sdk and soroban-token-sdk.

We only introduced it in order to be able to run the doctest in
migration guide soroban-sdk. I've moved the migration guide to the
respective crate (soroban-token-sdk). This may slightly reduce
discoverability, but I've left a reference to it in the main SDK
migration guide.

### Why

It's apparently impossible to publish crates with circular dependency
between them, even the dependency only occurs in dev-dependencies (or at
least I wasn't able to quickly find a way around that).

### Known limitations

N/A