Filter out events for the failed contract calls in `events().all()` test utility (#1565)
### What

Filter out events for the failed contract calls in `events().all()` test
utility

### Why

Due to an unfortunate design choice host returns all the events from the
`events()` call, including diagnostic events and contract events from
the failed calls. While this design may be revisited going forward, this
is a quick fix to bring the `events().all()` utility up to the
expectation of returning the contract events that would be emitted by
the call.

### Known limitations

N/A