Add soroban-meta package for wasm meta parsing (#1575)
### What
  Add soroban-meta package with wasm meta parsing.

  ### Why
Mirror the soroban-spec crate providing the same functionality for
extracting meta from contract wasm files. I was writing a utility that
interacted with both contract specs and contract meta, and it was
convenient to be able to use the soroban-spec crate to extract the spec,
and I was having to rewrite the same logic for the meta. It's been rare,
but someone else asked me in the past how to extract meta and I had to
tell them to do it manually because this crate didn't exist. The crate
API surface area matches that of soroban-spec minus the base64 functions
because the base64 functions don't get any use.