Make test wasms build more consistently (#1576)
### What
Remove the rssdkver (Rust SDK Version) meta entry from the builds of the
test wasms in this repository.

And build the test wasms once with the minimum supported rust version
(msrv) and use that build for test runs.

  ### Why
When building contracts with the soroban-sdk imported, the Rust SDK
Version is embedded in a meta entry. For the test wasms in this
repository this makes their builds unstable, and inconsistent. Every
commit on the repository results in a slightly different build given
that the revision is included and is forever changing. This makes it
challenging to use the wasms in some places where the hash or the raw
bytes get captured, such as test snapshots, and soon expanded generated
code (#1572).

The simplest way to remove this problem is to remove the meta entry just
when building the local test vectors.

Each version of rust can also cause differences in the wasm build, so
for tests, only testing against the wasms built with the msrv keeps the
wasm binaries stable for the longest period of time. This change
required some refactoring of the ci process, to separate building and
testing. That refactor also moved us away from using a separate set of
commands in ci vs locally, now ci just runs make commands.

The local Makefile is updated to build the wasms with the msrv as well
for convenience.