Various fixups to build (#1580)
### What
Install the msrv ahead of using it as an override. Remove making
warnings declined in the Makefile. Do not rebuild test wasms when
running tests. Change the default target to just run tests.

  ### Why
  Few things I wish I had done slightly differently in #1576.

Warnings are enabled in the makefile, but that needs to be set only in
CI, and only on the msrv now that the CI uses the makefile. Right now
warnings are enabled for both msrv and latest, when they shouldn't be.
Having warnings as errors off on local is good.

The tests run by the makefile always build the test wasms, but in CI
that doesn't need to happen as the test wasms are injected as an
artifact.

The default target currently calls the check target, which no longer
exists.