Track changes to generated code in test vectors (#1572)
### What
Add a job that expands the generated code of each test wasm both when
built locally for the host with tests, and when built for wasm, and
verifies if the committed files are up-to-date.

  ### Why
So that PRs show diffs when there are changes to generated code. To
catch unexpected and make hyper-visible the impact of changes to macros.

There's always an argument with this type of files we're committing that
it will get too noisy, too cumbersome, and that it won't be the right
balance and tradeoff in maintainence cost. This is an experiment to some
degree and I think we'll find out during Q4 if it pays off or not and
then we can retro, because we have changes planned that'll change the
generated code (e.g. #1507 #1535) and we'll see how it goes. We can
decide to remove this if the tradeoff doesn't end up being right.

For #1544

Dependent on:
- https://github.com/stellar/binaries/pull/45
- https://github.com/stellar/rs-soroban-sdk/pull/1580