Fix undefined behaviour miri warning in sdk testutils (#1586)
### What
Update `ctor` dependency from `0.2.9` to `0.5.0`.

  ### Why
It was reported in #1582 that undefined behaviour was detected by Miri.
The undefined behaviour was in the `ctor` crate, but was fixed in
https://github.com/mmastrac/rust-ctor/pull/343 which was released in
`ctor` `0.5.0`.

### Note

I tried adding miri to CI, but it takes too long, more than an hour.
I've left the makefile target for running miri for future convenience.

Fix #1582