Remove all permissions in test with soroban-examples workflow (#1594)
### What
Remove all permissions in the test with soroban-examples workflow file.

  ### Why
This explicitly declares that the workflow requires no GitHub token
permissions, following security best practices by applying the principle
of least privilege. The workflow pulls code from another repo and builds
it. That other repo is an internal @stellar repo, but it has a different
set of permissions and access and so nothing in that repo should be able
to escalate into this repo.